+++ src/Sulu/Component/Webspace/Document/Initializer/WebspaceInitializer.php
@@ -109,7 +109,8 @@ class WebspaceInitializer implements InitializerInterface
 
             $persistOptions = ['ignore_required' => true];
             if (!$homeDocument) {
-                $homeDocument = new HomeDocument();
+                /** @var HomeDocument $homeDocument */
+                $homeDocument = $this->documentManager->create('home');
                 $persistOptions['path'] = $homePath;
                 $persistOptions['auto_create'] = true;
             } else {